peer: abandon cancelled calls if the peer does not reply
When the context of a Call ends, Call sends a cancel packet to the peer.
Ordinarily the peer will cancel the handler and send a reply confirming
cancellation. But if the peer is unable to cancel the pending operation, that
reply may not arrive for a long time.

To avoid blocking the caller, set a watchdog timer that will complete then
pending call with an error if the peer does not reply in a timely manner.