peer: do not re-use request IDs for unacknowledged cancellations
In 110c89c we made it so a calling peer will (eventually) return control to the
host after the context expires, even if the remote peer does not acknowledge
the cancellation. However, when this occurs we should not yield the request ID
for reuse, since the remote peer may still claim it as active.