peer: add a Clone method
This method makes a new unstarted peer from an existing peer, in which the
method and packet handlers of the original peer are preserved. The peers are
independent after cloning, but this simplifies setup for multiple peers on the
same service.

Also fix a bug in the handling of base context functions discovered while
testing this.