spec,peer: clarify handling of duplicate requests
Require that when a duplicate request ID is received, it terminates both the
new request and the pending request. Allowing the existing request to continue
is confusing: If the caller tracks both, we have no way to know which one will
get the error, and the other will quietly succeed.

Since this is a weird corner anwyay, just fail both requests.