peer: mark local calls and handle them specially
When a Handler is called via Exec, from the local peer, mark the context of
that call so that invocations of Call from within that Handler are dispatched
directly into the target handler (provided one exists) rather than forwarded to
the remote peer.  Update the tests to properly exercise this.

Rationale: By design, handlers can call back into the calling peer, which is
usually the remote. But when you run Exec, the calling peer is the local one,
and to call back successfully it would have to know whether the call originated
remotely. Rather than bake this reasoning into the Handler, make it so Call in
a Handler will take care of deciphering who originated the invocation.