peers: use mnet.Network so tests can run in synctest
Drop the fake network fixtures and use mnet.Nework throughout.  This allows the
Loop test to use the synctest managed clock instead of having to really wait.
With that, also increase the number of calls spammed, and distribute the waits
with a random number to give the race detector more to work with.